migrate components, stylesheets, and snippets to Content
