+++ lib/beacon/content/component.ex
@@ -1,5 +1,15 @@
-defmodule Beacon.Components.Component do
-  @moduledoc false
+defmodule Beacon.Content.Component do
+  @moduledoc """
+  Components
+
+  > #### Do not create or edit components manually {: .warning}
+  >
+  > Use the public functions in `Beacon.Content` instead.
+  > The functions in that module guarantee that all dependencies
+  > are created correctly and all processes are updated.
+  > Manipulating data manually will most likely result
+  > in inconsistent behavior and crashes.
+  """
 
   use Ecto.Schema
   import Ecto.Changeset
