+++ lib/beacon/content/snippets/parser.ex
@@ -1,0 +1,5 @@
+defmodule Beacon.Content.Snippets.Parser do
+  @moduledoc false
+
+  use Solid.Parser.Base, custom_tags: [Beacon.Content.Snippets.TagHelper]
+end
