+++ lib/beacon/loader/component_module_loader.ex
@@ -3,6 +3,7 @@ defmodule Beacon.Loader.ComponentModuleLoader do
 
   require Logger
 
+  alias Beacon.Content.Component
   alias Beacon.Loader
 
   def load_components(_site, [] = _components) do
@@ -30,7 +31,7 @@ defmodule Beacon.Loader.ComponentModuleLoader do
     end
   end
 
-  defp render_component(%Beacon.Components.Component{site: site, name: name, body: body}) do
+  defp render_component(%Component{site: site, name: name, body: body}) do
     file = "site-#{site}-component-#{name}"
     ast = Beacon.Template.HEEx.compile_heex_template!(file, body)
