--- lib/beacon/snippets/parser.ex
@@ -1,5 +1,0 @@
-defmodule Beacon.Snippets.Parser do
-  @moduledoc false
-
-  use Solid.Parser.Base, custom_tags: [Beacon.Snippets.Tags.Helper]
-end
